+++ protaskinate/entities/__init__.py
@@ -1,2 +1,4 @@
 from .user import User
 from .task import Task
+from .project import Project
+from .comment import Comment