+++ protaskinate/repositories/__init__.py
@@ -1,2 +1,4 @@
 from .user_repository import user_repository
 from .task_repository import task_repository
+from .project_repository import project_repository
+from .comment_repository import comment_repository