+++ protaskinate/services/__init__.py
@@ -1,2 +1,4 @@
 from .user_service import user_service
 from .task_service import task_service
+from .project_service import project_service
+from .comment_service import comment_service