+++ protaskinate/services/task_service.py
@@ -1,6 +1,6 @@
 """protaskinate/services/task_service.py"""
 
-from typing import List, Literal
+from typing import List, Optional
 
 from protaskinate.entities import Task
 from protaskinate.repositories import task_repository
@@ -8,18 +8,27 @@ from protaskinate.repositories import task_repository
 
 class TaskService:
     """Class representing a service for tasks"""
-    def get_all(self,
-                order_by_fields: list[Literal["title", "created_at", "priority"]] | None = None,
-                reverse: list[bool] | None = None) -> List[Task]:
-        """Get all tasks"""
-        return task_repository.get_all(order_by_fields, reverse)
+    def get_all_by_project(self,
+                project_id: int,
+                order_by_fields: Optional[List[str]],
+                reverse: Optional[List[bool]]) -> List[Task]:
+        """Get all tasks by project"""
+        return task_repository.get_all({"project_id": project_id}, order_by_fields, reverse)
 
-    def update(self, task_id: int, **kwargs):
+    def get_by_id_and_project_with_comments(self, task_id: int, project_id: int) -> Optional[Task]:
+        """Get task by ID and project with comments"""
+        return task_repository.get_join_comments({"id": task_id, "project_id": project_id})
+
+    def update(self, task_id: int, project_id: int, **kwargs) -> Optional[Task]:
         """Update the task"""
-        task_repository.update(task_id, **kwargs)
+        return task_repository.update({"id":task_id, "project_id": project_id}, kwargs)
 
-    def create(self, **kwargs):
+    def create(self, **kwargs) -> Optional[Task]:
         """Create a task"""
-        task_repository.create(**kwargs)
+        return task_repository.create(kwargs)
+
+    def delete(self, task_id: int, project_id: int) -> None:
+        """Delete a task"""
+        return task_repository.delete({"id": task_id, "project_id": project_id})
 
 task_service = TaskService()