+++ docs/changelog.md
@@ -4,6 +4,28 @@ ProTaskinate changelog
 
 This format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [v0.2.0-beta] - 2024-10-01
+
+### Added
+
+- User is able to register
+- User is able to create a task with additional fields: assignee, deadline, and description
+- User is able to view all tasks with the additional fields: assignee, deadline, and description
+- User is able to additionally update the assignee of a task
+- User is able to delete a task
+- User is able to view a task's details by clicking on it
+- User is able to view all projects and their tasks
+- User is able to delete a project
+- User is able to view a tasks comments
+- User is able to add a comment to a task
+
+### Fixed
+
+- User can view active page in navbar
+- Protaskinate has a more consistent look and feel
+- Production database should no longer be flaky
+- Improved error handling
+
 ## [v0.1.0-beta] - 2024-09-17
 
 ### Added